Fixing several memory leaks in MultipleTermPositions
In the constructor a TermPositions array tps is created, but not
deleted. The constructor TermPositionsQueue don't take the
ownership of tps. Instead this array is copied.

skipTo() will pop TermPositions from the _termPositionsQueue, i.e.
remove them. If this TermPosition isn't putted back to the queue,
then the TermPosition will be closed, but not deleted.

The same in close() and in the destructor.